Upgrade, 1.2r5X -> 1.2r6X.
Pick up PREFIX and X11BASE from the environment if it is set there.
(That way you can build it correctly with a different PREFIX/X11BASE
from the top, or you can go in there and do a "make" too.)

Remove RESTRICTED, but this one IS_INTERACTIVE now.  We decided to shove
the license in the user's face and ask them if they will promise to read
it first. :)