Update to new DEC SRC release 3.6.
Use new custom distfiles which are MUCH smaller than the ones from
DEC.

Use the gcc-2.7.2.1 sources on the system if they are found in
/usr/src/contrib, to avoid having to fetch that distfile.

Use an existing Modula-3 compiler to bootstrap the new one, if
there happens to be one installed on the system.  Again, this
eliminates the need to fetch one of the distfiles.

Update the built-in thread-safe malloc to the latest version of
phkmalloc.