Make the depended file be consistent with what net/p5-IO installs.
Note, I have no idea if this is the correct fix or not, but I haven't
heard back from the maintainer or any of the millions of perl hackers.
At least it is not worse than what we have now (build net/p5-IO always,
because the depended file does not exist).