Upgrade to v0.37
99% of this update was from the original PR, which upgraded this port to 0.35
Because I sat on this for a week (sorry) another version was released.

I did not include the ports version number in the DESCR file as patched in
the PR, however, because I feel it is not needed.

Also the patches below have been removed because the author included a
a variation of them in the source code.

PR:		ports/11053
Submitted by:	Gianmarco Giovanelli <gmarco@giovannelli.it>