Fix a "USET" -> "USE" again.  IPv6 should work for this port.  Again.
Go to a much more convenient scheme for distfiles/ignorefiles.  There
will be a lot less change from now on... the release name not being
embedded in them helps a lot.

Fix an unquoted "${CVS_DATE}" so cvs update isn't always run when
we're in one of the first 9 days of a month in CVS_DATE.

Update to OpenSSH-1.2.2, which doesn't really mean anything since there
are no source releases anyway...

The port has been verified to work with pdksh 5.2.14 as /bin/sh, and
about 7 times faster.