Add SDL port of id Software's famous Quake game. Unfortunately, due to SDL's requirement of linuxthreads, this port will only work on systems running FreeBSD 4.0 and later. :-(
But, I would just like to rant about how cool sdlquake is. It's so FAST
(heck, takes about 90 seconds to compile+link on my PII-300), it's so
beautiful, and it's so much fun!! :-)

Okay, I know the Q3A freaks are gonna look at me funny, but hey, I'm free
of Windows (*grumble* except for scientific apps), and they're not. ;-)