Add WITH_UPDATES install option.
When the webmin version changes on the master site, patches for previous
versions may be removed (they are not stored in a version dependent
subdirectory). This means that the port of version A will break as soon as
version B is released, because the required updates disappear.

If the updates are optional, webmin users still get a chance to install the
port without updating their ports collection. It is also easier to support
users because, by default, they will run the same "official" version of
webmin.

PORTREVISION bump: if you reinstall without WITH_UPDATES you will get an
unpatched version of webmin.